Use tf.CustomCallbacks rather than redefining it
