+++ discojs/discojs-core/src/models/gpt/layers.ts
@@ -1,10 +1,11 @@
 import * as tf from '@tensorflow/tfjs'
+import type { GPTConfig } from './config.js'
 import type { ModelSize } from './config.js'
 
 /**
  * Defines a range, from 0 to T, that is used to create positional embeddings
  */
-export class Range extends tf.layers.Layer {
+class Range extends tf.layers.Layer {
   static readonly className = 'Range'
 
   computeOutputShape (inputShape: tf.Shape | tf.Shape[]): tf.Shape | tf.Shape[] {
@@ -26,7 +27,7 @@ export class Range extends tf.layers.Layer {
 }
 tf.serialization.registerClass(Range)
 
-export class LogLayer extends tf.layers.Layer {
+class LogLayer extends tf.layers.Layer {
   static readonly className = 'LogLayer'
 
   computeOutputShape (inputShape: tf.Shape | tf.Shape[]): tf.Shape | tf.Shape[] {
@@ -50,7 +51,7 @@ type CausalSelfAttentionConfig =
     & Record<'blockSize' | 'nHead' | 'nEmbd' | 'dropout', number>
     & { bias: boolean }
 
-export class CausalSelfAttention extends tf.layers.Layer {
+class CausalSelfAttention extends tf.layers.Layer {
   static readonly className = 'CausalSelfAttention'
 
   private readonly nHead: number
@@ -232,7 +233,7 @@ tf.serialization.registerClass(GELU)
 
 type MLPConfig = Required<ModelSize> & Record<'blockSize' | 'residDrop', number>
 
-export function MLP (config: MLPConfig): tf.LayersModel {
+function MLP (config: MLPConfig): tf.LayersModel {
   return tf.sequential({ layers: [
     tf.layers.dense({
       name: 'mlp/c_fc',
@@ -256,7 +257,7 @@ export function MLP (config: MLPConfig): tf.LayersModel {
 
 type BlockConfig = CausalSelfAttentionConfig & MLPConfig & { debug: boolean }
 
-export function TransformerBlock (conf: BlockConfig): tf.LayersModel {
+function TransformerBlock (conf: BlockConfig): tf.LayersModel {
   const config = Object.assign({ name: 'h' }, conf)
   const inputs = tf.input({ shape: [config.blockSize, config.nEmbd] })
   let x1, x2
@@ -283,4 +284,76 @@ export function TransformerBlock (conf: BlockConfig): tf.LayersModel {
   x2 = tf.layers.add().apply([x1 as tf.SymbolicTensor, x2 as tf.SymbolicTensor])
 
   return tf.model({ name: config.name, inputs, outputs: x2 as tf.SymbolicTensor })
+}
+
+
+/**
+ * The GPTArchitecture specifically defines a GPT forward pass, i.e.,
+ * what are the inputs, the successive transformer blocks and the outputs. It is then 
+ * used to create a GPTModel
+ * 
+ * @param conf GPTConfig
+ * @returns model, tf.LayersModel, which supports model(inputs), model.predict and model.apply
+ */
+export function GPTArchitecture (config: Required<GPTConfig>): tf.LayersModel {
+  const inputs = tf.input({ shape: [null] })
+
+  //Token embedding
+  const tokEmb = config.tokEmb
+    ? tf.layers.embedding({
+      name: config.name + '/wte',
+      inputDim: config.vocabSize,
+      outputDim: config.nEmbd,
+      embeddingsInitializer: 'zeros',
+      embeddingsRegularizer: undefined,
+      activityRegularizer: undefined
+    }).apply(inputs) as tf.SymbolicTensor
+    : inputs
+
+  // Positional embedding
+  const range = new Range({}).apply(inputs)
+  let posEmb = tf.layers.embedding({
+    name: config.name + '/wpe',
+    inputDim: config.blockSize,
+    outputDim: config.nEmbd,
+    embeddingsInitializer: 'zeros'
+  }).apply(range) as tf.SymbolicTensor
+  
+  if (config.debug) {
+    posEmb = new LogLayer({ name: 'posEmb' }).apply(posEmb) as tf.SymbolicTensor
+  }
+
+  // token and positional embeddings are added together
+  let x = tf.layers.add().apply([tokEmb, posEmb])
+  //dropout
+  x = tf.layers.dropout({name: 'drop', rate: config.embdDrop}).apply(x)
+  if (config.debug) {
+    x = new LogLayer({ name: 'dropadd' }).apply(x)
+  }
+
+  //Apply successively transformer blocks, attention and dense layers
+  for (let i = 0; i < config.nLayer; i++) {
+    x = TransformerBlock(
+      Object.assign({}, config, { name: config.name + '/h/' + i })
+    ).apply(x)
+  }
+  // Normalization
+  x = tf.layers.layerNormalization({ name: config.name + '/ln_f', epsilon: 1e-5 })
+    .apply(x)
+  if (config.debug) {
+    x = new LogLayer({ name: 'fin/ln' }).apply(x)
+  }
+
+  // Append a language modeling head if specified
+  if (config.lmHead) {
+    x = tf.layers.dense({
+      name: 'lm_head',
+      units: config.vocabSize,
+      inputDim: config.nEmbd,
+      inputShape: [config.blockSize, config.nEmbd],
+      useBias: false
+    }).apply(x)
+  }
+
+  return tf.model({ inputs, outputs: x as tf.SymbolicTensor })
 }
